Remove excess variant files.
From now on, we want customers to duplicate the "ARDUINO_NANO33BLE" folder of the arduino board package, rename it to "SF_MM_nRF52840_PB", and only copy over the MM specific pin definition files. This is so that the MM nRF variant is using the most recent/compatible variant files.